Throw FailureErrors when a test fails
The custom errors are designed to look reasonable as presented in Mocha
under Node. No guarantees at this point about what they'll look like in
browsers.

They have a .inputs value, which is an array of the inputs to the
property which caused it to fail. This is not currently documented nor
exposed to the user when using gentest under Mocha, which should change.