Fix leak of unused structure in shrink tree
While not a complete fix for #11, this does prevent the pathological
property described in #11 from resulting in an out-of-memory crash.
(Instead, that property now just spins indefinitely while exploring a
way-too-big shrink tree, with no built-in limits. Full fix will involve
adding limits to how hard we'll work to shrink.)